Add liquibase.headless true
This will avoid the prompt:

Do you want to see this operation's report in Liquibase Hub, which improves team collaboration?
If so, enter your email. If not, enter [N] to no longer be prompted, or [S] to skip for now, but ask again next time (default "S"):

(...) Which is not good for automated scripts.