chore(create-release.yml): add step to build and push Docker image to Docker Hub
The previous version of the workflow file did not include a step to build and push the Docker image to Docker Hub. This commit adds a new step to the workflow that uses the `docker/build-push-action` action to build and push the Docker image. The action is configured to use the current directory as the build context, enable caching, and push the image to Docker Hub. The action is also configured to build the image for both `linux/amd64` and `linux/arm64` platforms. The image is tagged with multiple tags, including `latest`, the extension version, and the minor version.