fix(test.yml): update the command to check if the desired log string is present in the container logs
The previous command was checking the container logs using `docker logs $CONTAINER_NAME`, but it was not capturing the output of the `liquibase update` command. This caused the check to fail even if the update was successful. The updated command `docker exec $CONTAINER_NAME liquibase update --defaultsFile=/liquibase/changelog/liquibase.properties --changelog-file=/changelog/example-changelog.xml 2>&1 | grep -q "$LOG_STRING"` captures the output of the `liquibase update` command and checks if the desired log string is present.