chore: add .DS_Store files to .gitignore
feat(test.yml): add test for extension loading in the GitHub Actions workflow

The .DS_Store files are binary files generated by macOS that are not necessary for version control. Adding them to the .gitignore file will prevent them from being tracked.

In the test.yml file, a new step is added to test the loading of an extension in the GitHub Actions workflow. It starts a Docker container, runs a Liquibase command, and checks if the desired log string is present. This test ensures that the extension loading functionality is working correctly.