chore: remove unused JAR files and update liquibase-mssql.properties
- Remove mssql-jdbc-8.2.2.jre11.jar and liquibase-redshift-4.25.0.jar as they are no longer needed
- Update liquibase-mssql.properties to remove the classpath for mssql-jdbc-8.2.2.jre11.jar

fix(test.yml): fix file paths in docker commands and increase sleep time

- Fix the file path for the changelog-file in the docker commands in the test.yml file
- Increase the sleep time from 10 to 20 seconds in the docker run command

feat(test.yml): add support for liquibase-redshift extension and mssql driver

- Add support for loading the liquibase-redshift extension in the Test extension loading step
- Add support for the mssql driver in the Test driver connection step
- Stop and remove the container before adding the mssql driver in the Test driver connection step

chore(test.yml): update custom entrypoint command

- Update the custom entrypoint command in the Test custom entrypoint step to use the correct script name

The changes were made to improve the functionality and maintainability of the code. The unused JAR files were removed to reduce clutter and potential conflicts. The liquibase-mssql.properties file was updated to remove the classpath for the removed JAR file. The file paths in the docker commands in the test.yml file were fixed to ensure proper execution. The sleep time in the docker run command was increased to allow for sufficient initialization time. Support for the liquibase-redshift extension and mssql driver was added to enhance the testing capabilities. The container was stopped and removed before adding the mssql driver to avoid conflicts. The custom entrypoint command was updated to use the correct script name for execution.