feat(liquibase.docker.properties): add liquibase.headless property to enable headless mode for Liquibase
The liquibase.docker.properties file was added with the following content:
```
liquibase.headless: true
```
This property is set to true to enable headless mode for Liquibase. Headless mode allows Liquibase to run without any user interface, which is useful for running Liquibase in a Docker container or in a non-interactive environment.