fix: allow container to run with arbitrary user IDs in OpenShift
Fixes #367

Changes the /liquibase directory ownership from liquibase:liquibase to
liquibase:root and adds group read/execute permissions (g+rx).

This enables the container to run in OpenShift and other Kubernetes
environments that use random UIDs with GID 0 (root group) for security.

Without this fix, containers fail to start when run with non-default
user IDs because the /liquibase directory is not accessible.